提交info参数为空的bug
提交info参数为空的bug

+++ Art/Assets/Toolkit/BundleChain/Editor/BuildWindow/AssetInfos.cs
@@ -16,7 +16,7 @@ namespace lhFramework.Tools.Bundle
         [ProtoMember(4)]
         public string hash { get; set; }
         [ProtoMember(5)]
-        public float size { get; set; }
+        public long size { get; set; }
     }
     [ProtoContract]
     public class AssetInfos
